create Common IT package
