+++ integration-tests/common/src/main/java/com/okta/spring/tests/common/code/BasicRedirectCodeFlowApplication.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.okta.spring.tests.oauth2.spring.code;
+package com.okta.spring.tests.common.code;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
