+++ integration-tests/common/src/main/resources/application.yml
