+++ pom.xml
@@ -42,6 +42,7 @@
         <module>oauth2</module>
         <module>sdk</module>
         <module>okta-spring-boot-starter</module>
+        <module>integration-tests/common</module>
         <module>integration-tests/oauth2</module>
         <module>integration-tests/oauth2-reactive</module>
         <module>integration-tests/oauth2-spring</module>
@@ -88,6 +89,11 @@
             <dependency>
                 <groupId>com.okta.spring</groupId>
                 <artifactId>okta-spring-boot-integration-tests-oauth2</artifactId>
+                <version>2.0.2-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>com.okta.spring</groupId>
+                <artifactId>okta-spring-boot-integration-tests-common</artifactId>
                 <version>2.0.2-SNAPSHOT</version>
             </dependency>
